+++ examples/WebClientRepeating_RP2040_SPI1/defines.h
@@ -16,7 +16,7 @@
 #define _ETG_LOGLEVEL_                      2
 
 // Default to use W5100. Must change to false for W5500, W5100S, for faster SPI clock
-#define USE_W5100                           true
+//#define USE_W5100                           true
 
 #define USING_SPI2                          true