+++ examples/WebClient_ESP_SPI2/defines.h
@@ -20,7 +20,7 @@
 #define _ETG_LOGLEVEL_                      2
 
 // Default to use W5100. Must change to false for W5500, W5100S, for faster SPI clock
-#define USE_W5100                           true
+//#define USE_W5100                           true
 
 // For ESP32
 #define ETHERNET_USE_ESP32