+++ src/Ethernet_Generic.h
@@ -32,7 +32,7 @@
   OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
   SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
   
-  Version: 2.5.1
+  Version: 2.5.2
     
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -46,6 +46,7 @@
   2.4.1   K Hoang      25/08/2022 Auto-select SPI SS/CS pin according to board package
   2.5.0   K Hoang      26/08/2022 Using raw_address() as default only for arduino-pico for compatibility
   2.5.1   K Hoang      01/09/2022 Slow SPI clock for old W5100 shield using SAMD Zero
+  2.5.2   K Hoang      06/09/2022 Slow SPI clock only when necessary. Improve support for SAMD21
  *****************************************************************************************************************************/
 
 #pragma once
@@ -56,13 +57,13 @@
 
 /////////////////////////////////////////////////////////
 
-#define ETHERNET_GENERIC_VERSION            "Ethernet_Generic v2.5.1"
+#define ETHERNET_GENERIC_VERSION            "Ethernet_Generic v2.5.2"
 
 #define ETHERNET_GENERIC_VERSION_MAJOR      2
 #define ETHERNET_GENERIC_VERSION_MINOR      5
-#define ETHERNET_GENERIC_VERSION_PATCH      1
+#define ETHERNET_GENERIC_VERSION_PATCH      2
 
-#define ETHERNET_GENERIC_VERSION_INT        2005001
+#define ETHERNET_GENERIC_VERSION_INT        2005002
 
 /////////////////////////////////////////////////////////